explicit access control - part 4 + moving into separate files
